refactor: move waterPlot to its own file
